+++ extensions/solusvm-2/CHANGELOG.md
@@ -1,3 +1,9 @@
 # SolusVM 2 Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- View Projects
+- Update Project (name, description)
+- View Servers in your Projects
+- Update Server (name, description)
+- Start, Stop Server